oms-monolith/requirements-otel.txt
# OpenTelemetry Dependencies - Unified Version Set
# Core API and SDK (Latest stable version)
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0

# Exporters
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0

# Core Instrumentation packages (0.45b0 for compatibility)
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-instrumentation-asyncio==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-httpx==0.45b0

# Propagators
opentelemetry-propagator-b3==1.24.0
opentelemetry-propagator-jaeger==1.21.0

# Semantic conventions
opentelemetry-semantic-conventions==0.45b0

# Utils
opentelemetry-util-http==0.45b0

# Optional instrumentation packages (not all versions exist)
# opentelemetry-instrumentation-aiohttp
# opentelemetry-instrumentation-asyncpg
# opentelemetry-instrumentation-pymongo
# opentelemetry-instrumentation-sqlalchemy
# opentelemetry-instrumentation-celery